backend/requirements.txt

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pydantic>=2.7.4
python-multipart>=0.0.9
email-validator>=2.1.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
resend>=0.8.0
slowapi>=0.1.9
supabase>=2.0.0
aiohttp>=3.9.0
